Rearrange it all
