--- README.md
@@ -1,4 +1,0 @@
-# s3strm/ffprobe
-
-s3 event -> lambda -> ffprobe.json
-
