--- bin/fire_for_ids
@@ -1,15 +1,0 @@
-#!/usr/bin/env bash
-STACK_NAME="s3strm-ffprobe"
-LAMBDA_ARN=$(
-  aws cloudformation describe-stacks \
-    --stack-name "${STACK_NAME}" \
-    --query 'Stacks[].Outputs[?OutputKey==`LambdaArn`].OutputValue' \
-    --output "text"
-)
-
-payload="$($(dirname $0)/sample_event $@)"
-aws lambda invoke \
-  --function-name "${LAMBDA_ARN}" \
-  --invocation-type "Event" \
-  --payload "${payload}" \
-  /dev/null
