--- bin/get_setting
@@ -1,10 +1,0 @@
-#!/usr/bin/env bash
-file=$(dirname $0)/../settings.inc
-source ${file}
-
-if [[ -z ${!1} ]]; then
-  aws s3 cp s3://${GENERAL_BUCKET}/settings.inc ${file} &> /dev/null
-  source ${file}
-fi
-
-echo ${!1}
