--- bin/get_stack_output
@@ -1,16 +1,0 @@
-#!/usr/bin/env bash
-STACK_NAME=$1
-OUTPUT_NAME=$2
-
-if [[ $# -eq 2 ]]; then
-  query='Stacks[].Outputs[?OutputKey==`'${OUTPUT_NAME}'`].OutputValue[]'
-  output="text"
-else
-  query='Stacks[].Outputs[]'
-  output="json"
-fi
-
-aws cloudformation describe-stacks  \
-  --stack-name "${STACK_NAME}"      \
-  --query ${query}                  \
-  --output "${output}"
