--- cfn/ffprobe.yaml
@@ -1,70 +1,0 @@
-AWSTemplateFormatVersion: 2010-09-09
-Description : Get metadata from an mp4 file using ffprobe
-
-Parameters:
- CodeKey:
-   Description: The key of the Lambda function
-   Type: String
-
-Resources:
-  Lambda:
-    Type : AWS::Lambda::Function
-    Properties:
-      Code:
-        S3Bucket: !ImportValue s3strm-settings-bucket
-        S3Key: !Ref CodeKey
-      Description: Get the metadata from an mp4 file using ffprobe
-      FunctionName: s3strm-ffprobe
-      Handler: ffprobe.lambda_handler
-      MemorySize: 128
-      Role: !GetAtt Role.Arn
-      Runtime: python2.7
-      Timeout: 120
-    DependsOn: Role
-
-  LambdaInvokePermission:
-    Type: AWS::Lambda::Permission
-    Properties:
-      Action: lambda:InvokeFunction
-      Principal: sns.amazonaws.com
-      SourceArn: !Join [ ":", [ "arn:aws:sns", Ref: "AWS::Region", Ref: "AWS::AccountId", "*" ] ]
-      FunctionName: !GetAtt Lambda.Arn
-
-  Role:
-    Type: AWS::IAM::Role
-    Properties:
-      AssumeRolePolicyDocument:
-        Version : 2012-10-17
-        Statement:
-          -
-            Effect: Allow
-            Principal:
-              Service:
-                - "lambda.amazonaws.com"
-            Action: "sts:AssumeRole"
-      Path: "/"
-      Policies:
-        -
-          PolicyName: s3strm-ffprobe
-          PolicyDocument:
-            Statement:
-              -
-                Effect: "Allow"
-                Action:
-                  - "logs:CreateLogGroup"
-                  - "logs:CreateLogStream"
-                  - "logs:PutLogEvents"
-                Resource: "arn:aws:logs:*:*:*"
-              -
-                Effect: Allow
-                Action: "s3:*"
-                Resource: !Join [ "", [ "arn:aws:s3:::", !ImportValue s3strm-movies-bucket, "/*" ] ]
-              -
-                Effect: Allow
-                Action: "s3:*"
-                Resource: !Join [ "", [ "arn:aws:s3:::", !ImportValue s3strm-settings-bucket, "/*" ] ]
-
-Outputs:
-  LambdaArn:
-    Value: !GetAtt Lambda.Arn
-
