--- cfn/lambda/setup.cfg
@@ -1,2 +1,0 @@
-[install]
-prefix=
