+++ lambdas/ffprobe/bin/regenerate_all
@@ -1,0 +1,31 @@
+#!/usr/bin/env bash
+BUCKET=$(
+  aws cloudformation list-exports \
+    --query 'Exports[?Name==`s3strm-movies-bucket`].Value' \
+    --output text
+)
+
+SET_SIZE=30
+
+IMDB_IDS=$(
+  aws s3 ls s3://${BUCKET}/ \
+    | awk '/tt[0-9]{7}/ { gsub(/\//, "", $2); print $2 }'
+)
+
+SETS=()
+i=0
+for imdb_id in ${IMDB_IDS}; do
+  if [[ $i -eq ${SET_SIZE} ]]; then
+    i=0
+    SETS+=("${str}")
+    str=""
+  fi
+  i=$(( i + 1 ))
+  str+="${imdb_id} "
+done
+
+for set in "${SETS[@]}"; do
+  echo "Regenerating for ${set}"
+  $(dirname $0)/fire_for_ids ${set}
+  sleep 5
+done
