+++ lambdas/ffprobe/bin/sample_event
@@ -1,0 +1,71 @@
+#!/usr/bin/env bash
+if [[ $# -eq 0 ]]; then
+  echo "Usage: $0 <imdb_id>" >&2
+  exit 1
+fi
+
+BUCKET_NAME=$(
+  aws cloudformation list-exports                           \
+    --query 'Exports[?Name==`s3strm-movies-bucket`].Value'  \
+    --output text
+)
+
+message='{ "Records": [ '
+first=true
+for imdb_id in "$@"; do
+  key="${imdb_id}/video.mp4"
+
+  ${first} || message+=","
+  message+='
+    {
+     "eventVersion": "2.0",
+     "eventTime": "1970-01-01T00:00:00.000Z",
+     "requestParameters": {
+      "sourceIPAddress": "127.0.0.1"
+     },
+     "s3": {
+      "configurationId": "testConfigRule",
+      "object": {
+       "eTag": "0123456789abcdef0123456789abcdef",
+       "sequencer": "0A1B2C3D4E5F678901",
+       "key": "'${key}'",
+       "size": 1024
+      },
+      "bucket": {
+       "arn": "arn:aws:s3:::'${BUCKET_NAME}'",
+       "name": "'${BUCKET_NAME}'",
+       "ownerIdentity": {
+        "principalId": "EXAMPLE"
+       }
+      },
+      "s3SchemaVersion": "1.0"
+     },
+     "responseElements": {
+      "x-amz-id-2": "EXAMPLE123/5678abcdefghijklambdaisawesome/mnopqrstuvwxyzABCDEFGH",
+      "x-amz-request-id": "EXAMPLE123456789"
+     },
+     "awsRegion": "'${AWS_DEFAULT_REGION}'",
+     "eventName": "ObjectCreated:Put",
+     "userIdentity": {
+      "principalId": "EXAMPLE"
+     },
+     "eventSource": "aws:s3"
+    }
+  '
+  first=false
+done
+message+=']}'
+
+event='
+  {
+    "Records": [
+      {
+        "Sns": {
+        "Message": "'$(echo ${message} | sed 's/\"/\\\"/g')'"
+        }
+      }
+    ]
+  }
+'
+
+echo ${event}
