+++ lambdas/ffprobe/src/setup.cfg
@@ -1,0 +1,2 @@
+[install]
+prefix=
